In this stage, you need to add checks for the following five errors to your program:
[S002] Indentation is not a multiple of four;

[S003] Unnecessary semicolon after a statement (note that semicolons are acceptable in comments);

[S004] Less than two spaces before inline comments;

[S005] TODO found (in comments only and case-insensitive);

[S006] More than two blank lines preceding a code line (applies to the first non-empty line).

Please note that:

If a line contains the same stylistic issue several times, your program should print the information only once. However, if a single line has several issues with different types of error codes, they should be printed as a sorted list.

To simplify the task, we consider it acceptable if your program finds some false-positive stylistic issues in strings, especially in multi-lines ('''...''' and """...""").

We recommend that you break your code into a set of functions to avoid confusion.